Upgrade Spring to 3.1 and Hibernate to 3.6
In addition to updating many POMs to support the more modular Spring and
Hibernate packaging, a series of small changes are needed for the
upgrade. This upgrade is to Spring 3.1.4 and Hibernate 3.6.10.

 * Clean up javassist and hibernate-jmx issues
 * Fix up cm-authz tests and clean up POMs for HSQLDB tests
 * Hibernate dependency upgrades in kernel
   * Add hibernate-commons-annotations
   * Add hibernate-jpa-2.0-api
 * Update RSF to 0.8.0-SNAPSHOT in master POM
 * Remove gradebook indexes now autocreated by Hibernate

The additional Hibernate jars need to be declared because, even though
hibernate-core typically packs them, we mark it as provided for the
modules that inherit it. This causes the transitive dependencies to also
be marked provided, so they need to be declared explicitly in our
kernel/deploy/shared POM to be available to everything.