Fix Hibernate DTDs and convert text fields to materialized_clob
The materialized_clob type is new and removes the ambiguity of how long
textual fields are mapped. With materialized_clob, there is a CLOB
column and it maps as a Java String. This is opposed to earlier versions
where the text type could correspond to different SQL types depending on
length and vendor. Now, the text type maps a LONGVARCHAR to a String,
and with recent versions of Oracle (at least 10g and 11g), CLOB is
strictly preferred. This change will require some before/after testing
and potentially a conversion script to change any remaining LONG or
LONGVARCHAR columns to CLOBs. MySQL behavior should remain unchanged as
the text types have always been handled as CLOBs (e.g., as TEXT or
MEDIUMTEXT).

Hibernate now expects its DTD URLs to be hibernate.org as opposed to
the older sourceforge.net URLs. These have all been changed.