SAK-23733 - Fix gradebook assignment visibility
This change adds a solution to a problem in the grouping/visibility
support for external gradebook assignments. It accommodates tools that
place items in the gradebook but do not implement an
ExternalAssignmentProvider, falling back to the behavior that was
present in the 2.8 release. That is, when no provider specifies whether
a particular item should be included or excluded in a student's view and
grade calculation, it should be included.

This behavior is accomplished by comparing the sets of all gradebook
items for a site, all items known by all providers, and all items that
providers indicate should be visible. It adds a temporary interface so
there is no API breakage. The Assignments and Samigo providers implement
the single new method. The Lessons implementation is expected shortly.

Note, though, that any existing providers that do not implement the new
getAllExternalAssignments method will not be able to suppress items in
the gradebook because of the algorithm used (see
GradebookExternalAssessmentServiceImpl.java:523). This is not expected
to be a problem, since there are no known providers that implement
group/visibility controls outside the core, though an adequate window
should be provided before the 2.9.3 release so contrib tools can be
verified and implement the interface if they choose (which is
recommended).